Add user login and fetch functionality in Home
The changes introduced in this commit encompass both a user login mechanism and a mechanism to fetch users. This has been done utilizing a combination of React hooks useEffect and useState to control state and lifecycle of the components. The user login is done via a POST request and the response token is stored in local storage, then being set as local state, triggering fetchUsers function. The users are then fetched by making a GET request, and the users data is then displayed.