Add Tailwind CSS configuration and update user login.
This commit includes the addition of postcss and tailwind configuration files, which are necessary for utilizing Tailwind CSS in the project. Dependencies including postcss, tailwindcss, and autoprefixer have also been added to package.json for the same reason.

Furthermore, changes have been made to the user login feature. Paths have been specified for different types of files in tailwind.config.js. The home page now fetches a list of users and provides a link to edit and delete a user.

An error message is now displayed on the login page when a login error occurs. CSS imports for the global styles have been added as well, together with new bootstrap dependencies and react-router-dom for navigation.

Inclusive, an Edit page for users profiles was added, providing information loading feedback, contributing to a better user experience.

Contrary to what was done, the console.log(token) statement was added to src/app/users.js, but it should be avoided to include executable code in commit diffs. Moreover, it potentially exposes sensitive information.